refactor: initial commit

+++ src/plugin.ts
@@ -1,7 +1,7 @@
 import fs from "fs";
 import path from "path";
 import { PostHTML } from "posthtml";
-import { createHash, hashFileName, processFile } from "./utils";
+import hasha from "hasha";
 
 const DEFAULT_PATH = "";
 const DEFAULT_HASH_LENGTH = 20;
@@ -11,6 +11,23 @@ const DEFAULT_OPTIONS: IOptions = {
   css: true,
   js: true,
 };
+
+const REGEX_HASH = new RegExp(/\[hash.*]/g);
+
+export function replaceHash(str: string, buffer: Buffer | string) {
+  const match = str.match(REGEX_HASH);
+
+  if (match == null) {
+    return str;
+  }
+
+  const [_, len] = match[0].replace(/\[|]/g, "").split(":");
+
+  return str.replace(
+    REGEX_HASH,
+    hasha(buffer).slice(0, Number(len) || DEFAULT_HASH_LENGTH)
+  );
+}
 
 function plugin(options = DEFAULT_OPTIONS) {
   return function posthtmlHash(tree: PostHTML.Node) {
@@ -39,10 +56,9 @@ function plugin(options = DEFAULT_OPTIONS) {
       const pathToFile = options.path || "";
       const file = path.join(process.cwd(), pathToFile, fileName);
 
-      processFile(file, () => {
+      if (fs.existsSync(file)) {
         const buffer = fs.readFileSync(file);
-        const hash = createHash(buffer, hashLength);
-        const hashedFileName = hashFileName(fileName, hash);
+        const hashedFileName = replaceHash(fileName, buffer);
         const hashedFile = path.join(process.cwd(), pathToFile, hashedFileName);
 
         fs.renameSync(file, hashedFile);
@@ -52,7 +68,7 @@ function plugin(options = DEFAULT_OPTIONS) {
         } else if (attrs.src) {
           node.attrs!.src = hashedFileName;
         }
-      });
+      }
 
       return node;
     });
